refactor(money): last Double money fields to BigDecimal — slice 5, migration closed
The four columns still on DOUBLE PRECISION after V24/V25:
  - bundles.discount_percentage        (percentage)
  - promo_campaigns.discount_percent   (percentage)
  - subscription_plans.amount          (amount)
  - return_requests.refund_amount      (amount)

V30 converts all four to NUMERIC(12,2), ROUND(...) in every USING clause as in
V24/V25. V25 already settled that percentage columns travel as NUMERIC(12,2) too
(it did products.discount), so 12.50% is now storable as 12.50.

Two seams close with the conversion:
  - ReturnServiceImpl dropped asLegacyDouble(BigDecimal) and the double-narrowing
    getOrderTotal — the refund is copied straight from the order total, a
    BigDecimal, as it always should have been.
  - SubscriptionServiceImpl sent Stripe (long)(amount * 100); now
    Money.of(amount).toCents(), and the plan-changed check is compareTo != 0
    rather than a scale-sensitive BigDecimal.equals.

Money gains a percentage(BigDecimal) overload (the double one delegates to it) so
the discount rate no longer round-trips through double into the arithmetic.

DTOs, GdprExportData.ReturnRequestEntry updated to match. No frontend or GraphQL
change: the schema already maps every migrated BigDecimal money field to the
Float scalar, and Jackson serialises BigDecimal as a JSON number.

Tests: LegacyMoneyFieldsNumericTest (column types + exact round-trips);
existing ReturnServiceTest / PromoCampaignSweepTest / RepositoryNPlusOneAuditTest
adjusted for the type. Full suite 447 green.

+++ ecom-backend/src/main/java/com/ecommerce/project/payload/PromoCampaignDTO.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -18,7 +19,7 @@ public class PromoCampaignDTO {
     @NotNull
     @DecimalMin("0.0")
     @DecimalMax("100.0")
-    private Double discountPercent;
+    private BigDecimal discountPercent;
     // LocalDateTime.parse(null, ...) throws NPE inside the service.
     @NotBlank
     private String startTime;
